Module trellis::core

Structs

ButtonEvent
Describes a button event that occurred. At the moment only button press is implemented.

LedButton
A particular button and LED combination in the grid. Inconsistently used in the trellis implementation at the moment.

Trellis
The Trellis device abstraction.

Enums

Col
A column in the trellis grid. See the project readme for a scheme of the trellis orientation.

Row
A row in the trellis grid. See the project readme for a scheme of the trellis orientation.

Type Definitions

EventLoopHandler
The callback type for the button_evt_loop (see the Trellis struct for details).